properly compute days until

+++ slack.go
@@ -116,9 +116,9 @@ func (s *SlackSvc) GetMessages() []string {
 		if err != nil {
 			continue
 		}
-		daysUntil := d.Sub(time.Now()).Hours() / 24
+		daysUntil := int64(d.Sub(time.Now()).Hours() / 24)
 		if daysUntil > 0 {
-			msgs = append(msgs, fmt.Sprintf("%s is *%f* days away!", v, daysUntil))
+			msgs = append(msgs, fmt.Sprintf("%s is *%d* days away!", v, daysUntil))
 		}
 	}
